Break setter fn generation into its own module
