+++ lib/parse-route-string.js
@@ -14,15 +14,12 @@ function logFull( obj ) {
 
 var CollectionRequest = require( './shared/collection-request' );
 var inherit = require( 'util' ).inherits;
+var generatePathPartSetter = require( './generate-path-part-setter' );
 
 // All valid routes in API v2 beta 11
 var routes = require( './endpoint-response.json' ).routes;
 
-function getValues( obj ) {
-	return typeof obj === 'object' ? Object.keys( obj ).map(function( key ) {
-		return obj[ key ];
-	}) : [];
-}
+var getValues = require( 'lodash' ).values;
 
 /*
 
@@ -136,63 +133,16 @@ Object.keys( routesByNamespace ).forEach(function( namespace, idx ) {
 		 * // @return {isLeaf} A boolean indicating whether the processed node is a leaf
 		 */
 		function extractSetterFromNode( node ) {
-			var dynamicChildren = getValues( node.children ).filter(function( childNode ) {
-				return childNode.namedGroup === true;
-			});
+			var setterFn;
 
 			// For each node, add its handler to the relevant "level" representation
 			handler.addLevelOption( node.level, _.pick( node, 'validate', 'methods' ) );
 
-			var setterFn;
-			if ( node.namedGroup ) {
-				/**
-				 * Set a dymanic (named-group) path part of a query URL.
-				 *
-				 * @chainable
-				 * @param  {String|Number} val The path part value to set
-				 * @return {Object} The handler instance (for chaining)
-				 */
-				setterFn = function( val ) {
-					/* jshint validthis:true */
-					this.setPathPart( node.level, val );
-					return this;
-				};
-			} else {
-				/**
-				 * Set a non-dymanic (non-named-group) path part of a query URL, and
-				 * set the value of a subresource if an input value is provided and
-				 * exactly one named-group child node exists.
-				 *
-				 * @example
-				 *
-				 *     // revisions() is a non-dynamic path part setter:
-				 *     wp.posts().id( 4 ).revisions();       // Get posts/4/revisions
-				 *     wp.posts().id( 4 ).revisions( 1372 ); // Get posts/4/revisions/1372
-				 *
-				 * @chainable
-				 * @param  {String|Number} [val] The path part value to set (if provided)
-				 *                               for a subresource within this resource
-				 * @return {Object} The handler instance (for chaining)
-				 */
-				setterFn = function( val ) {
-					/* jshint validthis:true */
-					// If the path part is not a namedGroup, it should have exactly one
-					// entry in the names array: use that as the value for this setter,
-					// as it will usually correspond to a collection endpoint.
-					this.setPathPart( node.level, node.names[ 0 ] );
-
-					// If this node has exactly one dynamic child, this method may act as
-					// a setter for that child node
-					var dynamicChild = dynamicChildren.length === 1 && dynamicChildren[ 0 ];
-					if ( typeof val !== 'undefined' && dynamicChild ) {
-						this.setPathPart( dynamicChild.level, val );
-					}
-					return this;
-				};
-			}
-
-			// First level is set implicitly
+			// First level is set implicitly, no dedicated setter needed
 			if ( node.level > 0 ) {
+
+				setterFn = generatePathPartSetter( node );
+
 				node.names.forEach(function( name ) {
 					// camel-case the setter name
 					var name = name
